Moved non-callback related functions into separate files to clean up the main application script file.
